Tidy up
Move register conversion functions to utils
s/to_the/pow
Index consistency
